feat: refactoring structure
